+++ src/mod/am/am.scala
@@ -4,8 +4,8 @@ import dds.trig._
 import utils._
 import chisel3._
 
-class AM[T <: Data](gen: T)(carrierFreq: Long, phaseDelta: Int, lutWidth: Int, clkFreq: Long)(
-  implicit analog: Analog[T]
+class AM[T <: Data](gen: T)(carrierFreq: Int, phaseDelta: Int, lutWidth: Int, clkFreq: Int)(implicit
+  analog: Analog[T]
 ) extends Module {
   override def desiredName = s"am_cf$carrierFreq"
   val io                   = IO(new SISO(gen)).suggestName("AM")
