+++ src/mod/psk/psk.scala
@@ -10,7 +10,7 @@ class BPSKIO[T <: Data](gen: T) extends Bundle {
   val out  = Output(gen)
 }
 
-class BPSK[T <: Data](gen: T)(carrierFreq: Long, phaseWidth: Int, lutWidth: Int, clkFreq: Long)(
+class BPSK[T <: Data](gen: T)(carrierFreq: Int, phaseWidth: Int, lutWidth: Int, clkFreq: Int)(
   implicit analog: Analog[T]
 ) extends Module {
   override def desiredName = s"bpsk_cf$carrierFreq"
@@ -26,7 +26,7 @@ class QPSKIO[T <: Data](gen: T) extends Bundle {
   val out  = Output(gen)
 }
 
-class QPSK[T <: Data](gen: T)(carrierFreq: Long, phaseWidth: Int, lutWidth: Int, clkFreq: Long)(
+class QPSK[T <: Data](gen: T)(carrierFreq: Int, phaseWidth: Int, lutWidth: Int, clkFreq: Int)(
   implicit analog: Analog[T]
 ) extends Module {
   override def desiredName = s"qpsk_cf$carrierFreq"
@@ -54,11 +54,11 @@ class nPSKIO[T <: Data](gen: T)(dw: Int) extends Bundle {
 }
 
 class nPSK[T <: Data](gen: T)(
-  carrierFreq: Long,
+  carrierFreq: Int,
   dw: Int,
   phaseWidth: Int,
   lutWidth: Int,
-  clkFreq: Long
+  clkFreq: Int
 )(implicit analog: Analog[T])
     extends Module {
   override def desiredName = s"psk_n${dw}_cf$carrierFreq"
